Tree Functionality Implemented
Now a player can create more trees by putting 2 trees next to eachother, tree's can only do this twice before they are destroyed, this implementation will require polish as it can be a bit unpredictable, but it works and so it's good enough for the first prototype.

Also fixed some minor bugs with the gaint, so now it doesn't pick up child objects, but rather its parent.