Control changes
The Giant now rotates by pressing a and d instead of strafing.

The Giant moves forwards and backwards depending on it's own rotation, instead of the direction the camera is facing.

Created an animation for turning left, need to do the same for turning right, as well as implementing them into gameplay