Overhaul on picking up
Objects can now be picked up from further away

The object to pick up is now decided on distance from the giant, not what entered the range detector last

Fixed glitch where the last object in the detector could be picked up even if it was no longer in range

Pickup detector now fetches the root of the collisions gameobject.

Fixed glitch where anything inside the area that a villager wanted to build was hard to pick up due to the detector trying to pick up the area display.

No more errors from the pick-up variable not being assigned to as if there is nothing to pick up, the variable will be filled by the directional light (As it is so far out of the way it won't intervine with anything, and is also impossible to pick up)