Menu, Destruction, Water
Added a Main Menu with limited functionality atm

Made a universal destruction animation which also gives the giant a tear

Animated the water, added groundwater beyond the terrain and made another plane which give the illusion of a water effect when the camera is submerged.

Fixed new villagers not having the center point defined.