Started Work on saving
Changed villagers so they no longer destroy themselves and create a clone when founding a village. Instead the one in the foreign place can no longer build for the rest of the day.

Finished work on "Save Data" file. Need to finish "Save" file and start working on Load file (Also creating a loadmap for bideford).

Issue of founding villages not working if starting an action has not been worked on.